// src/shared/interfaces/influencer-subjects.interface.ts
import { IInfluencer } from './influencer.interface';
export interface IInfluencerSubjects {
    idInfluencer: IInfluencer;
    description: string;
    subject: string;
    subject_value1: number;
    subject2: string;
    subject_value2: number;
    subject3: string;
    subject_value3: number;
    subject4: string;
    subject_value4: number;
    subject5: string;
    subject_value5: number;
}